stream and store in consensus
