+++ server/v2/cometbft/go.mod
@@ -1,6 +1,6 @@
 module cosmossdk.io/server/v2/cometbft
 
-go 1.21.1
+go 1.21
 
 replace (
 	cosmossdk.io/core => ../../../core
@@ -13,6 +13,7 @@ replace (
 )
 
 require (
+	cosmossdk.io/api v0.7.2
 	cosmossdk.io/core v0.12.0
 	cosmossdk.io/errors v1.0.0
 	cosmossdk.io/log v1.2.1
@@ -39,6 +40,7 @@ require (
 	github.com/cockroachdb/tokenbucket v0.0.0-20230807174530-cc333fc44b06 // indirect
 	github.com/cometbft/cometbft-db v0.9.1 // indirect
 	github.com/cosmos/cosmos-db v1.0.0 // indirect
+	github.com/cosmos/cosmos-proto v1.0.0-beta.3 // indirect
 	github.com/cosmos/ics23/go v0.10.0 // indirect
 	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
 	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.2.0 // indirect
