+++ server/v2/cometbft/server.go
@@ -1,125 +1,102 @@
 package cometbft
 
-import (
-	"context"
-	"fmt"
+// import (
+// 	"context"
+// 	"fmt"
 
-	"cosmossdk.io/log"
-	"cosmossdk.io/server/v2/appmanager"
-	cometlog "cosmossdk.io/server/v2/cometbft/log"
-	"cosmossdk.io/server/v2/cometbft/types"
-	"cosmossdk.io/server/v2/core/transaction"
-	"cosmossdk.io/store/v2/snapshots"
-	abciserver "github.com/cometbft/cometbft/abci/server"
-	abci "github.com/cometbft/cometbft/abci/types"
-	cmtcfg "github.com/cometbft/cometbft/config"
-	"github.com/cometbft/cometbft/node"
-	"github.com/cometbft/cometbft/p2p"
-	pvm "github.com/cometbft/cometbft/privval"
-	"github.com/cometbft/cometbft/proxy"
-	// genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
-)
+// 	"cosmossdk.io/log"
+// 	"cosmossdk.io/server/v2/appmanager"
+// 	cometlog "cosmossdk.io/server/v2/cometbft/log"
+// 	"cosmossdk.io/server/v2/cometbft/types"
+// 	"cosmossdk.io/server/v2/core/transaction"
+// 	abciserver "github.com/cometbft/cometbft/abci/server"
+// 	abci "github.com/cometbft/cometbft/abci/types"
+// 	cmtcfg "github.com/cometbft/cometbft/config"
+// 	"github.com/cometbft/cometbft/node"
+// 	"github.com/cometbft/cometbft/p2p"
+// 	pvm "github.com/cometbft/cometbft/privval"
+// 	"github.com/cometbft/cometbft/proxy"
+// 	// genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
+// )
 
-type Config struct {
-	Name            string // TODO: we might want to put some of these in the app manager
-	Version         string
-	InitialHeight   uint64
-	MinRetainBlocks uint64
-	IndexEvents     map[string]struct{}
-	HaltHeight      uint64
-	HaltTime        uint64
+// type CometBFTServer struct {
+// 	Node   *node.Node
+// 	app    abci.Application
+// 	logger log.Logger
 
-	SnapshotManager *snapshots.Manager
+// 	config    Config
+// 	cleanupFn func()
+// }
 
-	AddrPeerFilter types.PeerFilter // filter peers by address and port
-	IdPeerFilter   types.PeerFilter // filter peers by node ID
+// func NewCometBFTServer[T transaction.Tx](logger log.Logger, app appmanager.AppManager[T], cfg Config, voteExtHandler types.VoteExtensionsHandler) *CometBFTServer {
+// 	logger = logger.With("module", "cometbft-server")
+// 	return &CometBFTServer{
+// 		logger: logger,
+// 		app:    NewConsensus[T](app, cfg),
+// 		config: cfg,
+// 	}
+// }
 
-	Transport  string
-	Addr       string
-	Standalone bool
-	Trace      bool
+// func (s *CometBFTServer) Start(ctx context.Context) error {
+// 	wrappedLogger := cometlog.CometLoggerWrapper{Logger: s.logger}
+// 	if s.config.Standalone {
+// 		svr, err := abciserver.NewServer(s.config.Addr, s.config.Transport, s.app)
+// 		if err != nil {
+// 			return fmt.Errorf("error creating listener: %w", err)
+// 		}
 
-	CmtConfig *cmtcfg.Config
-}
+// 		svr.SetLogger(wrappedLogger)
 
-type CometBFTServer struct {
-	Node   *node.Node
-	app    abci.Application
-	logger log.Logger
+// 		return svr.Start()
+// 	}
 
-	config    Config
-	cleanupFn func()
-}
+// 	nodeKey, err := p2p.LoadOrGenNodeKey(s.config.CmtConfig.NodeKeyFile())
+// 	if err != nil {
+// 		return err
+// 	}
 
-func NewCometBFTServer[T transaction.Tx](logger log.Logger, app appmanager.AppManager[T], cfg Config, voteExtHandler types.VoteExtensionsHandler) *CometBFTServer {
-	logger = logger.With("module", "cometbft-server")
-	return &CometBFTServer{
-		logger: logger,
-		app:    NewConsensus[T](app, cfg),
-		config: cfg,
-	}
-}
+// 	s.Node, err = node.NewNode(
+// 		ctx,
+// 		s.config.CmtConfig,
+// 		pvm.LoadOrGenFilePV(s.config.CmtConfig.PrivValidatorKeyFile(), s.config.CmtConfig.PrivValidatorStateFile()),
+// 		nodeKey,
+// 		proxy.NewLocalClientCreator(s.app),
+// 		getGenDocProvider(s.config.CmtConfig),
+// 		cmtcfg.DefaultDBProvider,
+// 		node.DefaultMetricsProvider(s.config.CmtConfig.Instrumentation),
+// 		wrappedLogger,
+// 	)
+// 	if err != nil {
+// 		return err
+// 	}
 
-func (s *CometBFTServer) Start(ctx context.Context) error {
-	wrappedLogger := cometlog.CometLoggerWrapper{Logger: s.logger}
-	if s.config.Standalone {
-		svr, err := abciserver.NewServer(s.config.Addr, s.config.Transport, s.app)
-		if err != nil {
-			return fmt.Errorf("error creating listener: %w", err)
-		}
+// 	s.cleanupFn = func() {
+// 		if s.Node != nil && s.Node.IsRunning() {
+// 			_ = s.Node.Stop()
+// 		}
+// 	}
 
-		svr.SetLogger(wrappedLogger)
+// 	return s.Node.Start()
+// }
 
-		return svr.Start()
-	}
+// func (s *CometBFTServer) Stop() error {
+// 	defer s.cleanupFn()
+// 	if s.Node != nil {
+// 		return s.Node.Stop()
+// 	}
+// 	return nil
+// }
 
-	nodeKey, err := p2p.LoadOrGenNodeKey(s.config.CmtConfig.NodeKeyFile())
-	if err != nil {
-		return err
-	}
+// // returns a function which returns the genesis doc from the genesis file.
+// func getGenDocProvider(cfg *cmtcfg.Config) func() (node.ChecksummedGenesisDoc, error) {
+// 	return func() (node.ChecksummedGenesisDoc, error) {
+// 		//TODO: re-add this after fixing deps
+// 		// appGenesis, err := genutiltypes.AppGenesisFromFile(cfg.GenesisFile())
+// 		// if err != nil {
+// 		// 	return nil, err
+// 		// }
 
-	s.Node, err = node.NewNode(
-		ctx,
-		s.config.CmtConfig,
-		pvm.LoadOrGenFilePV(s.config.CmtConfig.PrivValidatorKeyFile(), s.config.CmtConfig.PrivValidatorStateFile()),
-		nodeKey,
-		proxy.NewLocalClientCreator(s.app),
-		getGenDocProvider(s.config.CmtConfig),
-		cmtcfg.DefaultDBProvider,
-		node.DefaultMetricsProvider(s.config.CmtConfig.Instrumentation),
-		wrappedLogger,
-	)
-	if err != nil {
-		return err
-	}
-
-	s.cleanupFn = func() {
-		if s.Node != nil && s.Node.IsRunning() {
-			_ = s.Node.Stop()
-		}
-	}
-
-	return s.Node.Start()
-}
-
-func (s *CometBFTServer) Stop() error {
-	defer s.cleanupFn()
-	if s.Node != nil {
-		return s.Node.Stop()
-	}
-	return nil
-}
-
-// returns a function which returns the genesis doc from the genesis file.
-func getGenDocProvider(cfg *cmtcfg.Config) func() (node.ChecksummedGenesisDoc, error) {
-	return func() (node.ChecksummedGenesisDoc, error) {
-		//TODO: re-add this after fixing deps
-		// appGenesis, err := genutiltypes.AppGenesisFromFile(cfg.GenesisFile())
-		// if err != nil {
-		// 	return nil, err
-		// }
-
-		// return appGenesis.ToGenesisDoc()
-		return node.ChecksummedGenesisDoc{}, nil
-	}
-}
+// 		// return appGenesis.ToGenesisDoc()
+// 		return node.ChecksummedGenesisDoc{}, nil
+// 	}
+// }
